Create Implementation_Strassen’s Matrix Multiplication.cpp
This is an implementation of Strassen's matrix multiplication in C++. The Strassen’s method of matrix multiplication is a typical divide and conquer algorithm. This algorithm is slightly faster than the general matrix multiplication algorithm. The general algorithm’s time complexity is O(n^3), while the Strassen’s algorithm is O(n^2.80). Please see (http://www.stoimen.com/blog/2012/11/26/computer-algorithms-strassens-matrix-multiplication/).

This code calculate the multiplication of any n*m matrix A by any m*p matrix B. If the matrices A, B are not of type 2^n x 2^n we fill the missing rows and columns with zeros (zero padding).